skin-disease-api/requirements.txt
flask==2.0.1
tensorflow==2.13.0
google-cloud-storage==2.5.0
pillow==9.0.1